+++ src/strings.hpp
@@ -32,7 +32,7 @@ bool str_starts_with(std::string const * const, std::string const * const);
  * @param to to index (exclusive)
  * @return if the substring inside the string contains no spaces aka only a "real world".
  */
-bool str_contains_single_word_from_to(std::string const *const, unsigned from, unsigned to);
+bool str_contains_single_word_from_to(std::string const *const, size_t from, size_t to);
 
 /**
  * Checks if string only contains given chars in given range.
@@ -42,7 +42,7 @@ bool str_contains_single_word_from_to(std::string const *const, unsigned from, u
  * @param chars
  * @return if string only contains given chars in given range
  */
-bool str_contains_only_given_chars(std::string const *const input, unsigned from, unsigned to, std::string const * const chars);
+bool str_contains_only_given_chars(std::string const *const input, size_t from, size_t to, std::string const * const chars);
 
 /**
  * Splits a string into sub strings by a string delimiter.
